Fixed all compiler warnings
