+++ adagucserverEC/CImgWarpNearestNeighbour.h
@@ -57,6 +57,7 @@ public:
   float legendLogAsLog;
   CDataSource * dataSource;
   CDrawImage *drawImage;
+  bool debug;
   //size_t prev_imgpointer;
   void init(CDataSource *dataSource,CDrawImage *drawImage,int tileWidth,int tileHeight){
     this->dataSource = dataSource;
@@ -76,6 +77,14 @@ public:
     if(dataSource->dfBBOX[2]<dataSource->dfBBOX[0]){
       dfSourceBBOX[0]=dataSource->dfBBOX[2];
       dfSourceBBOX[2]=dataSource->dfBBOX[0];
+    }
+    
+    debug = false;
+    
+    if(dataSource->cfgLayer->TileSettings.size()==1){
+      if(dataSource->cfgLayer->TileSettings[0]->attr.debug.equals("true")){
+        debug = true;
+      }
     }
     
     CStyleConfiguration *styleConfiguration = dataSource->getStyle();
@@ -228,7 +237,7 @@ public:
                         val = 5;
                       }
 #endif             */       
-                                    
+                 
                   
                   isNodata=false;
                   if(hasNodataValue){if(val==nodataValue)isNodata=true;}if(!(val==val))isNodata=true;
@@ -245,6 +254,19 @@ public:
                     
                     //drawImage->setPixelIndexed(dstpixel_x,dstpixel_y,drawImage->colors[(unsigned char)val]);
                     drawImage->setPixelIndexed(dstpixel_x,dstpixel_y,pcolorind);
+                  }
+                  if(debug){
+                    bool draw = false;
+                    bool draw2 = false;
+                    if(srcpixel_x ==0||srcpixel_x==width-1||srcpixel_y ==0||srcpixel_y==height-1){draw=true;}
+                    if((srcpixel_x ==10||srcpixel_x==width-10)&& srcpixel_y >10 &&srcpixel_y<height-10){draw2=true;}
+                    if((srcpixel_y ==10||srcpixel_y==width-10)&& srcpixel_x >10 &&srcpixel_x<width-10){draw2=true;}
+                    if(draw){
+                      drawImage->setPixelIndexed(dstpixel_x,dstpixel_y,249);
+                    }
+                    if(draw2){
+                      drawImage->setPixelIndexed(dstpixel_x,dstpixel_y,244);
+                    }
                   }
                 }
               }
