+++ hclasses/CTypes.h
@@ -48,7 +48,7 @@
 
 #include "CTBaseType.h"
 #include "CTString.h"
-#include "CTStringRef.h"
+// #include "CTStringRef.h"
 //#define CTYPES_DEBUG
