+++ hclasses/CXMLParser.cpp
@@ -1,3 +1,4 @@
+
 /******************************************************************************
  * 
  * Project:  Helper classes
@@ -143,11 +144,10 @@ void CXMLParser::XMLElement::parse_element_names(void *_a_node,int depth){
   for (cur_node = a_node; cur_node; cur_node = cur_node->next) {
     if (cur_node->type == XML_ELEMENT_NODE) {
       char *content=NULL;
-      if(cur_node->children!=NULL)
-        if(cur_node->children->content!=NULL)
-          if(cur_node->children->type==XML_TEXT_NODE)
+      if(cur_node->children!=NULL){
+        if(cur_node->children->content!=NULL){
+          if(cur_node->children->type==XML_TEXT_NODE){
             content=(char*)cur_node->children->content;
-          {
             XMLElement child;
             child.name=(char*)cur_node->name;
             child.value=(char*)content;
@@ -157,6 +157,8 @@ void CXMLParser::XMLElement::parse_element_names(void *_a_node,int depth){
             }
             xmlElements.push_back(child);
           }
+        }
+      }
     }
   }
 }
@@ -192,7 +194,7 @@ CT::string CXMLParser::XMLElement::toXML(XMLElement el,int depth){
     data+=el.getValue();
     data+= CT::string("</")+el.name+">\n";
   }else{
-    for(int i=0;i<depth;i++)data+="  ";data+= CT::string("</")+el.name+">\n";
+    for(int i=0;i<depth;i++){data+="  ";};data+= CT::string("</")+el.name+">\n";
   }
   return data;
 }
